Some steps back
In order to get a first release the changes in this version went some
steps back. The whole Upgrade-System and its progress were deleted and
all 4 turrets, even the upgrades, are buyable in the Shop. Neverless
this Update includes some great addition.

Added BuildingEffect (A ParticalEffect used when buildings are build)
Added hoverColor, when not enough Money to build (it's red)