Bug Fixes
This Version fixes the Bugs which made the earlier Versions unplayable.
Now there's just one Bug that turns up by starting the Game.

In the MainMenu Script: test = FindGameObject<Toggle>
Although there is no Toggle until the Player enters the OptionsMenu.
The Bug needs to be adressed but doesnt make the Game unplayable.